requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
docutils==0.21.2
ecdsa==0.19.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
filetype==1.2.0
fonttools==4.57.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.25.0
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.1
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.46.2
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.3
wcwidth==0.2.13
yfinance==0.2.55
